pindah function untuk slider di halaman tersendiri

+++ functions.php
@@ -39,29 +39,7 @@ if (is_admin()) {
   });
 }
 
-function slider_post_init() {
-  $args = array(
-    'label' => 'Slider',
-    'public' => true,
-    'show_ui' => true,
-    'capability_type' => 'post',
-    'hierarchical' => false,
-    'rewrite' => array('slug' => 'slider'),
-    'query_var' => true,
-    'menu_icon' => 'dashicons-slides',
-    'menu_position' => 5,
-    'show_in_menu' => 'custom-post',
-    'supports' => array(
-      'title',
-      'thumbnail',
-      'author',
-      'page-attributes',
-    )
-  );
-  register_post_type( 'slider', $args );
-}
-add_action( 'init', 'slider_post_init' );
-
+include('functions/slider.php');
 include('functions/events.php');
 
 add_action( 'admin_menu', 'custom_post_menu' );
@@ -78,11 +56,13 @@ function wpt_save_meta( $post_id, $post ) {
     $meta['event_date'] = esc_textarea( $_POST['event_date'] );
     $meta['event_time'] = esc_textarea( $_POST['event_time'] );
   }
-  foreach ( $meta as $key => $value ) :
-    if ( 'revision' === $post->post_type ) return;
-    if ( get_post_meta( $post_id, $key, false ) ) update_post_meta( $post_id, $key, $value );
-    else add_post_meta( $post_id, $key, $value);
-    if ( ! $value ) delete_post_meta( $post_id, $key );
-  endforeach;
+  if ($meta) {
+    foreach ( $meta as $key => $value ) :
+      if ( 'revision' === $post->post_type ) return;
+      if ( get_post_meta( $post_id, $key, false ) ) update_post_meta( $post_id, $key, $value );
+      else add_post_meta( $post_id, $key, $value);
+      if ( ! $value ) delete_post_meta( $post_id, $key );
+    endforeach;
+  }
 }
 add_action( 'save_post', 'wpt_save_meta', 1, 2 );
